+++ sydtest/src/Test/Syd/Runner/Synchronous/Separate.hs
@@ -98,7 +98,7 @@ runSpecForestSynchronously settings testForest =
               )
       DefAroundAllWithNode func sdf -> do
         e <- ask
-        let HCons x _ = eExternalResources e
+        let outers = eExternalResources e
         liftIO $
           fmap SubForestNode
             <$> applySimpleWrapper
@@ -106,9 +106,9 @@ runSpecForestSynchronously settings testForest =
               ( \b ->
                   runReaderT
                     (goForest sdf)
-                    (e {eExternalResources = HCons b (eExternalResources e)})
+                    (e {eExternalResources = HCons b outers})
               )
-              x
+              outers
       DefAfterAllNode func sdf -> do
         e <- ask
         let externalResources = eExternalResources e